Update purchaseApi.yaml
Hi Attila,

We have made some minor adjustments to the API specification to make it more consistent overall.
The only changes that might cause issues are the model names. We have changed the following so we have a consistency in model (request) names:
•	PurchaseAuthorizationRequest -> AuthorizePurchaseRequest
•	PurchaseInitializationRequest -> InitializePurchaseRequest
•	CaptureRequest -> CapturePurchaseRequest
•	RefundRequest -> RefundPurchaseRequest
•	Termsandconditions -> getTermsAndConditions

Could you let us know if you agree with these changes? If not, we can always revert them to the previous values.
Furthermore we were wondering if you could give us the descriptions of the following models:
•	Ach
•	Eft
•	Bacs
These are part of the bank account (Account) model and a brief description could help the merchants understand more easily what information they need to provide.