Delete php-sdk.md
php temporary removed 
"Hi Attila, Isaac pushed an update for the documentation to fix the issues I've reported. Could you update these in Stoplight, and remove the php sdk one (we are not ready to release the php sdk yet)? And then we should be ready to publish on live."